Debut de l'édit qui marche

+++ src/java/com/stageo/controleur/DeleteOffreAction.java
@@ -5,7 +5,9 @@
  */
 package com.stageo.controleur;
 
+import com.stageo.beans.Avertissement;
 import com.stageo.beans.OffreStage;
+import com.stageo.beans.Utilisateur;
 import com.stageo.dao.OffreStageDAO;
 import com.stageo.singleton.Connexion;
 
@@ -17,15 +19,25 @@ public class DeleteOffreAction extends AbstractAction {
 
     @Override
     public String execute() {
-        try{       
+        try{
+            Utilisateur currentUser = (Utilisateur)request.getSession().getAttribute("utilisateur");
             OffreStageDAO offreDao = new OffreStageDAO(Connexion.getInstance());
-            OffreStage offreTemp = new OffreStage();
-            //SÉCURISÉ POUR TCHEKER QUE CEST LE BON USER !!!
-            offreTemp.setIdOffre(request.getParameter("id"));
-            offreDao.delete(offreTemp);
+            OffreStage offreTemp = offreDao.findById(request.getParameter("id"));
+            
+            if(offreTemp.getIdEmployeur().equals(currentUser.getIdUtilisateur())){
+                Avertissement aver = new Avertissement("L'offre à été supprimé.", "succes");
+                request.getSession().setAttribute("avertissement", aver);
+                offreDao.delete(offreTemp);
+            }
+            else{
+                Avertissement aver = new Avertissement("Vous ne pouvez pas supprimer des offres qui ne sont pas à vous.", "erreur");
+                request.getSession().setAttribute("avertissement", aver);
+            }
             return "gestionOffresStagesVueEmployeur";
         }catch(Exception e){
-            return "" + e;
+            Avertissement aver = new Avertissement(""+e, "erreur");
+            request.getSession().setAttribute("avertissement", aver);
+            return "gestionOffresStagesVueEmployeur";
         }
     }
 }
